MLaaS/walle/pyproject.toml
[tool.poetry]
name = "walle"
version = "0.1.0"
description = ""
authors = ["Henry Ruiz  <henryruiz22@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
google-cloud-vision = "^3.1.3"
matplotlib = "^3.6.1"
pillow = "^9.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
